move files out of project root and into e2e package
