--- e2e/constants.ts
@@ -1,1 +1,0 @@
-export const BASE_URL = process.env.E2E_ENDPOINT;
