+++ e2e/src/constants.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = process.env.E2E_ENDPOINT;
