+++ e2e/src/my-forms.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { MyForms } from '../packages/design/src/FormManager/routes';
+import { MyForms } from '../../packages/design/src/FormManager/routes';
 import { BASE_URL } from './constants';
 
 test('Go to MyForms', async ({ page }) => {
